requirements.txt
gradio>=4.44.1
transformers
torch
PyPDF2
python-docx
sentencepiece
protobuf 